backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
bcrypt
cryptography
pycryptodome
python-multipart
pillow
scipy